requirements.txt
# Core data science
numpy==1.24.3
pandas==2.0.3
scipy==1.10.1

# Machine learning
scikit-learn==1.3.0
imbalanced-learn==0.11.0
xgboost==1.7.6

# Deep learning
tensorflow==2.13.0
keras==2.13.1

# Explainable AI
shap==0.42.1
lime==0.2.0.1

# Web framework
flask==2.3.2
flask-cors==4.0.0
waitress==2.1.2

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Utilities
joblib==1.3.1
pyyaml==6.0
tqdm==4.65.0
python-dotenv==1.0.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Jupyter
jupyter==1.0.0
ipykernel==6.25.0